backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
dj-database-url==2.1.0
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0
kombu==5.3.4
